Vaadin v24へ更新

+++ src/main/java/jp/whitenoise/jftest/dao/魚種Dao.java
@@ -6,5 +6,5 @@ import org.springframework.stereotype.Repository;
 import jp.whitenoise.jftest.model.魚種;
 
 @Repository
-public interface 魚種Dao extends JpaRepository<魚種, Long> {
+public interface 魚種Dao extends JpaRepository<魚種, Integer> {
 }
